refactoring + improved doc
